fix: skip i18n format on falsy file row dates
Every file row called the i18n date formatter on render, even when the
row was missing both `updated_at` and `created_at`. twake-i18n's `f()`
calls date-fns `format()` which throws on falsy dates; the library
catches the throw but logs it via `console.error`, and our Sentry
config captures console errors as events. Synthetic rows in the
`/sharings` and shared-drives views (where entries can lack
timestamps) emit one event per row per render.

Extract the guarded format into a `useFormattedUpdatedAt` hook that
returns undefined for falsy dates, and use it at the three row sites
(File, GridFile, virtualized Cell). Mirror the pattern in the
virtualized LastUpdate cell so its title is only formatted when the
parent yielded a value. Drop the now-unused `f` and `isExtraLarge`
props from File and GridFile.
